Move anchor generation back to detector file
